+++ packages/ListBoxWithTags/stories/ListBoxWithTags.uncontrolled-users.stories.js
@@ -2,7 +2,7 @@ import React from "react";
 import { getStoryName } from "storybook/storyTree";
 import UncontrolledUsersApp from "./examples/UncontrolledUser";
 
-const storyName = getStoryName("ListBox");
+const storyName = getStoryName("ListBoxWithTags");
 
 export default {
   title: storyName,