+++ packages/ListBox/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Change
+
+- WithTags subcomponent was moved to their own packages no is part of @paprika/list-box-with-tags
+
 ## [1.0.0] - 2020-11-04
 
 ### Changed